+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fixed an issue where names of XSIAM content items were not normalized properly.
 * Fixed an issue where script yml files were not properly identified by `find_type`.
 * Added support for XSIAM Dashboard preview image.
+* Added support for Agent Config content item.
 
 ## 1.7.4